# Change daemon-side components to be fully `async` for performance improvements.