--- a/runhouse/servers/cluster_servlet.py
+++ b/runhouse/servers/cluster_servlet.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 
 class ClusterServlet:
-    def __init__(
+    async def __init__(
         self, cluster_config: Optional[Dict[str, Any]] = None, *args, **kwargs
     ):
 
@@ -30,22 +30,24 @@ class ClusterServlet:
     ##############################################
     # Cluster config state storage methods
     ##############################################
-    def get_cluster_config(self) -> Dict[str, Any]:
+    async def get_cluster_config(self) -> Dict[str, Any]:
         return self.cluster_config
 
-    def set_cluster_config(self, cluster_config: Dict[str, Any]):
+    async def set_cluster_config(self, cluster_config: Dict[str, Any]):
         self.cluster_config = cluster_config
 
-    def set_cluster_config_value(self, key: str, value: Any):
+    async def set_cluster_config_value(self, key: str, value: Any):
         self.cluster_config[key] = value
 
     ##############################################
     # Auth cache internal functions
     ##############################################
-    def add_user_to_auth_cache(self, token: str, refresh_cache: bool = True):
+    async def add_user_to_auth_cache(self, token: str, refresh_cache: bool = True):
         self._auth_cache.add_user(token, refresh_cache)
 
-    def resource_access_level(self, token: str, resource_uri: str) -> Union[str, None]:
+    async def resource_access_level(
+        self, token: str, resource_uri: str
+    ) -> Union[str, None]:
         # If the token in this request matches that of the owner of the cluster,
         # they have access to everything
         if configs.token and (
@@ -55,13 +57,13 @@ class ClusterServlet:
             return ResourceAccess.WRITE
         return self._auth_cache.lookup_access_level(token, resource_uri)
 
-    def user_resources(self, token: str) -> dict:
+    async def user_resources(self, token: str) -> dict:
         return self._auth_cache.get_user_resources(token)
 
-    def get_username(self, token: str) -> str:
+    async def get_username(self, token: str) -> str:
         return self._auth_cache.get_username(token)
 
-    def has_resource_access(self, token: str, resource_uri=None) -> bool:
+    async def has_resource_access(self, token: str, resource_uri=None) -> bool:
         """Checks whether user has read or write access to a given module saved on the cluster."""
         from runhouse.rns.utils.api import ResourceAccess
 
@@ -88,46 +90,46 @@ class ClusterServlet:
 
         return True
 
-    def clear_auth_cache(self, token: str = None):
+    async def clear_auth_cache(self, token: str = None):
         self._auth_cache.clear_cache(token)
 
     ##############################################
     # Key to servlet where it is stored mapping
     ##############################################
-    def mark_env_servlet_name_as_initialized(self, env_servlet_name: str):
+    async def mark_env_servlet_name_as_initialized(self, env_servlet_name: str):
         self._initialized_env_servlet_names.add(env_servlet_name)
 
-    def is_env_servlet_name_initialized(self, env_servlet_name: str) -> bool:
+    async def is_env_servlet_name_initialized(self, env_servlet_name: str) -> bool:
         return env_servlet_name in self._initialized_env_servlet_names
 
-    def get_all_initialized_env_servlet_names(self) -> Set[str]:
+    async def get_all_initialized_env_servlet_names(self) -> Set[str]:
         return self._initialized_env_servlet_names
 
-    def get_key_to_env_servlet_name_dict_keys(self) -> List[Any]:
+    async def get_key_to_env_servlet_name_dict_keys(self) -> List[Any]:
         return list(self._key_to_env_servlet_name.keys())
 
-    def get_key_to_env_servlet_name_dict(self) -> Dict[Any, str]:
+    async def get_key_to_env_servlet_name_dict(self) -> Dict[Any, str]:
         return self._key_to_env_servlet_name
 
-    def get_env_servlet_name_for_key(self, key: Any) -> str:
+    async def get_env_servlet_name_for_key(self, key: Any) -> str:
         return self._key_to_env_servlet_name.get(key, None)
 
-    def put_env_servlet_name_for_key(self, key: Any, env_servlet_name: str):
+    async def put_env_servlet_name_for_key(self, key: Any, env_servlet_name: str):
         if not self.is_env_servlet_name_initialized(env_servlet_name):
             raise ValueError(
                 f"Env servlet name {env_servlet_name} not initialized, and you tried to mark a resource as in it."
             )
         self._key_to_env_servlet_name[key] = env_servlet_name
 
-    def pop_env_servlet_name_for_key(self, key: Any, *args) -> str:
+    async def pop_env_servlet_name_for_key(self, key: Any, *args) -> str:
         # *args allows us to pass default or not
         return self._key_to_env_servlet_name.pop(key, *args)
 
-    def clear_key_to_env_servlet_name_dict(self):
+    async def clear_key_to_env_servlet_name_dict(self):
         self._key_to_env_servlet_name = {}
 
     ##############################################
     # Remove Env Servlet
     ##############################################
-    def remove_env_servlet_name(self, env_servlet_name: str):
+    async def remove_env_servlet_name(self, env_servlet_name: str):
         self._initialized_env_servlet_names.remove(env_servlet_name)